requirements.txt
copier~=9.2
dbrownell_DevTools
pytest-copie~=0.2
syrupy~=4.6
